updated code

+++ utilities/common.util.js
@@ -4,7 +4,9 @@ const config = require('../config/default.config');
 const ApplicationErrors = require('./application-errors');
 const CustomEmailError = require('./custom-error.util');
 const ajv = new Ajv({ allErrors: true, jsonPointers: true });
+const { Logger } = require('../utilities/logger.util')
 
+//TODO: check what ajv does
 require('ajv-errors')(ajv);
 
 
@@ -18,36 +20,38 @@ class CommonUtil {
 
   /**
   * This method will handle the failed messages.
-  * @param {Object} processedMessages
-  * @param {Object} records
   */
-  static async handleBatchFailure(recordsToDelete,failedMessageIds) {
+  static async handleBatchFailure(recordsToDelete, failedMessageIds) {
 
     const sucessMessageIds = [];
 
     const entriesToDelete = recordsToDelete.map((messageToDelete) => {
-      // Logger.getRecordLogger(messageToDelete.messageId).info('Started deleting record.');
+      Logger.getRecordLogger(messageToDelete.messageId).info('Started deleting record.');
       return {
         Id: messageToDelete.messageId,
         ReceiptHandle: messageToDelete.receiptHandle
       };
     });
+    
+    if (entriesToDelete.length) {
+      const deleteMsgRes = await sqs.deleteMessageBatch({
+        Entries: entriesToDelete,
+        QueueUrl: config.scoreQueueUrl
+      }).promise();
 
-    const deleteMsgRes = await sqs.deleteMessageBatch({
-      Entries: entriesToDelete,
-      QueueUrl: config.scoreQueueUrl
-    }).promise();
+      deleteMsgRes.Successful.forEach((record) => {
+        sucessMessageIds.push(record.Id);
+        Logger.getRecordLogger(record.Id).info('Successfully deleted record from queue.');
+      });
 
-    deleteMsgRes.Successful.forEach((record) => {
-      sucessMessageIds.push(record.Id);
-      // Logger.getRecordLogger(record.Id).info('Successfully deleted record from queue.');
-    });
+      deleteMsgRes.Failed.forEach((record) => {
+        failedMessageIds.push(record.Id);
+        Logger.getRecordLogger(record.Id).info('Failed to delete record from queue.');
+      });
+    }
 
-    deleteMsgRes.Failed.forEach((record) => {
-      failedMessageIds.push(record.Id);
-      // Logger.getRecordLogger(record.Id).info('Failed to delete record from queue.');
-    });
 
+    // need to throw here so that the all the batch does not vanish from SQS.
     throw new CustomEmailError(ApplicationErrors.BATCH_PROCESSING_FAILED, null, {
       info: {
         SuccessfulMessageIds: sucessMessageIds,
@@ -58,12 +62,11 @@ class CommonUtil {
 
   /**
   * This method will handle the failed messages.
-  * @param {Object} processedMessages
-  * @param {Object} recordData
-  * @param {Object} options
+  * if recordData has messageId, it will initialize logger with recordLevel data and use it for logging the error.
+  * if options has isFutherComputationsRequired set as true, it will not throw the error and only log it.
   */
   static handleError(error, recordData, options) {
-    // const logger = options.isRecordLevel ? Logger.getRecordLogger(recordData.messageId) : Logger.getLogger();
+    const logger = options.isRecordLevel ? Logger.getRecordLogger(recordData.messageId) : Logger.getLogger();
     let errCode;
     let errMsg;
 
@@ -93,7 +96,7 @@ class CommonUtil {
       };
       errMsg = JSON.stringify(errorObj);
     }
-    // logger.error({ err: error }, `Error Type= ${errCode}, Error= ${errMsg}`);
+    logger.error({ err: error }, `Error Type= ${errCode}, Error= ${errMsg}`);
     if (options.isFurtherComputationsNeeded) {
       return;
     }
